pacify clang
git-svn-id: https://svn.r-project.org/R/trunk@62314 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/modules/X11/dataentry.c
@@ -1320,7 +1320,7 @@ static void printstring(DEstruct DE, const char *ibuf, int buflen, int row, int 
                 wcs[j] = *(wcspc + j);
             wcs[j] = L'\0';
             w_p = wcs;
-            cnt = wcsrtombs(s, (const wchar_t **)&w_p, sizeof(s) - 1, NULL);
+            cnt = (int)wcsrtombs(s, (const wchar_t **)&w_p, sizeof(s) - 1, NULL);
             s[cnt] = '\0';
             if (textwidth(DE, s, (int)strlen(s)) < (bw - DE->text_offset))
                 break;
@@ -2689,7 +2689,7 @@ static void calc_pre_edit_pos(DEstruct DE)
     int i;
     int w;
 
-    xpoint.x = (int)DE->boxw[0];
+    xpoint.x = (short)DE->boxw[0];
     for (i = 1; i < DE->ccol; i++)
         xpoint.x += BOXW(DE->colmin + i - 1);
 #ifdef HAVE_XUTF8TEXTESCAPEMENT
@@ -2700,7 +2700,7 @@ static void calc_pre_edit_pos(DEstruct DE)
         w = XmbTextEscapement(font_set, buf, clength);
     xpoint.x += (w > BOXW(DE->colmin + DE->ccol - 1)) ? BOXW(DE->colmin + DE->ccol - 1) : w;
     xpoint.x += DE->text_offset;
-    xpoint.y = (int)(DE->hht + (DE->crow + 1) * DE->box_h - DE->text_offset);
+    xpoint.y = (short)(DE->hht + (DE->crow + 1) * DE->box_h - DE->text_offset);
 
     /*
       <FIXME>
